Adding an AuthenticatedController
By moving things into the ApplicationController, I was obliterating the
VisitorController's behavior.
