+++ app/controllers/application_controller.rb
@@ -27,35 +27,6 @@ class ApplicationController < ActionController::Base
   end
   helper_method :repository
 
-  # @todo With Cogitate this will need to be revisited
-  def authenticate_user!
-    authenticated_user = authenticate_with_http_basic do |group_name, group_api_key|
-      authorize_group_from_api_key(group_name: group_name, group_api_key: group_api_key)
-    end
-    if authenticated_user
-      @current_user = authenticated_user
-    else
-      super
-    end
-  end
-
-  # Required because the authorization layer is firing the current user test prior to the authenticate_user! action filter
-  # The end result was that the user for the web request came through as nil in the authorization layer.
-  #
-  # @todo With Cogitate this will need to be revisited
-  def current_user
-    return @current_user if @current_user
-    authenticate_user!
-    @current_user
-  end
-
-  # @todo With Cogitate this will need to be revisited
-  def authorize_group_from_api_key(group_name:, group_api_key:)
-    return false unless group_api_key
-    return false unless group_name
-    Sipity::Models::Group.find_by(name: group_name, api_key: group_api_key) || false
-  end
-
   private
 
   def message_for(key, options = {})
